week6/requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
numpy==2.1.1
opencv_python==4.10.0.84
Pillow==10.4.0
plotly==5.20.0
scipy==1.14.1
skimage==0.0
